backend/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
sqlalchemy==2.0.40
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
alembic==1.14.1
markdown==3.7
pillow==10.4.0
pydantic==2.10.6
pydantic-settings==2.2.1
pydantic[email]==2.10.6
psycopg2-binary==2.9.9  # PostgreSQL adapter
asyncpg==0.29.0  # Async PostgreSQL driver 